Port to Clang
The integrated assembler doesn't understand the `-Wa,-divide` option.
Thus we passed the `no-integrated-as` option to Clang.

Clang tries to use `%xmm` registers, which seems not working for xv6.
Thus we disabled SSE by passing `-mno-sse` to Clang.

This commit still has race condition.  When compiled with "-O2", the
kernel panics all the time; when compiled with "-O0", the kernel
panics quite often.  Hence this commit changes optimization level
from "-O2" to "-O0".